pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "mmexplorer"
version = "1.0.0"
description = "Fetch protein structure, apply missense mutation, energy-minimise, compare in browser"
authors = [{ name = "fredsanto (Federico Santoni)", email = "federico.santoni@gmail.com" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.10"
keywords = ["protein", "mutation", "bioinformatics", "openmm", "pdbfixer"]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]
dependencies = [
    "flask>=3.0",
    "biopython>=1.83",
    "openmm>=8.0",
    "pdbfixer>=1.9",
    "requests>=2.31",
    "numpy>=1.26",
]

[project.scripts]
protein-mutation = "app:main"

[tool.setuptools.packages.find]
where = ["."]

[tool.setuptools.package-data]
"*" = ["templates/**"]
